Demo_code/requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
boto3==1.17.58
botocore==1.20.112
cached-property==1.5.2
cachetools==4.2.2
celery==5.1.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
django-tenants==3.3.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
et-xmlfile==1.1.0
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==19.6.0
httpagentparser==1.9.1
httplib2==0.19.1
humanize==3.11.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
python3-linkedin==1.0.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
zipp==3.5.0
